backend/requirements.txt
bidict==0.23.1
blinker==1.8.2
click==8.1.7
dnspython==2.7.0
eventlet==0.37.0
Flask==3.0.3
Flask-Cors==5.0.0
Flask-SocketIO==5.4.1
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
h11==0.14.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
python-engineio==4.10.1
python-socketio==5.11.4
regex==2024.9.11
simple-websocket==1.1.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
Werkzeug==3.0.4
wsproto==1.2.0
GitPython==3.1.43
